added hook to hook files
